raytrace/requirements.txt
arcade==3.0
numpy==2.2.5
Pillow==11.2.1
